Merge pull request #1 from oczkers/master

+++ chromium/services/ui/ws/focus_controller.cc
@@ -5,6 +5,7 @@
 #include "services/ui/ws/focus_controller.h"
 
 #include "base/macros.h"
+#include "base/memory/ptr_util.h"
 #include "services/ui/public/interfaces/window_manager.mojom.h"
 #include "services/ui/ws/focus_controller_delegate.h"
 #include "services/ui/ws/focus_controller_observer.h"
@@ -185,9 +186,10 @@ bool FocusController::CanBeActivated(ServerWindow* window) const {
     bool is_minimized = false;
     const ServerWindow::Properties& props = window->properties();
     if (props.count(mojom::WindowManager::kShowState_Property)) {
+      // The type must match that of PropertyConverter::PrimitiveType.
       is_minimized =
           props.find(mojom::WindowManager::kShowState_Property)->second[0] ==
-          static_cast<int>(ui::mojom::ShowState::MINIMIZED);
+          static_cast<int64_t>(ui::mojom::ShowState::MINIMIZED);
     }
     if (!is_minimized)
       return false;
